resolve js

+++ ParseFloat.js
@@ -4,10 +4,10 @@ function onClickParseFloat(){
     document.getElementById("pResult").innerHTML=result ;
 }
    document.getElementById("pDemo").innerHTML = 
-  Number.parseInt("-10") + "<br>" +
-  Number.parseInt("-10.33") + "<br>" +
-  Number.parseInt("10") + "<br>" +
-  Number.parseInt("10.33") + "<br>" +
-  Number.parseInt("10 6") + "<br>" +  
-  Number.parseInt("10 years") + "<br>" +  
-  Number.parseInt("years 10");  
+  Number.parseFloat("-10") + "<br>" +
+  Number.parseFloat("-10.33") + "<br>" +
+  Number.parseFloat("10") + "<br>" +
+  Number.parseFloat("10.33") + "<br>" +
+  Number.parseFloat("10 6") + "<br>" +  
+  Number.parseFloat("10 years") + "<br>" +  
+  Number.parseFloat("years 10");
